Separate ob http handler
- #21
- Created ob http handler file.
- Renamed functions.
