+++ backend/src/http_handler.h
@@ -8,7 +8,15 @@
 #ifndef BACKEND_SRC_HTTP_HANDLER_H_
 #define BACKEND_SRC_HTTP_HANDLER_H_
 
+#define DEF_REG_UUID "[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}"
+
 bool init_http_handler(void);
 void term_http_handler(void);
+
+json_t* create_default_result(int code);
+
+void simple_response_error(evhtp_request_t *req, int status_code, int err_code, const char* err_msg);
+void simple_response_normal(evhtp_request_t *req, json_t* j_msg);
+
 
 #endif /* BACKEND_SRC_HTTP_HANDLER_H_ */
