+++ backend/src/ob_event_handler.h
@@ -8,7 +8,7 @@
 #ifndef SRC_EVENT_HANDLER_H_
 #define SRC_EVENT_HANDLER_H_
 
-int   run_outbound(void);
+bool  init_outbound(void);
 void  stop_outbound(void);
 
 #endif /* SRC_EVENT_HANDLER_H_ */
