+++ backend/src/ob_http_handler.h
@@ -1,0 +1,13 @@
+/*
+ * ob_http_handler.h
+ *
+ *  Created on: Mar 15, 2017
+ *      Author: pchero
+ */
+
+#ifndef BACKEND_SRC_OB_HTTP_HANDLER_H_
+#define BACKEND_SRC_OB_HTTP_HANDLER_H_
+
+bool init_ob_http_handler(void);
+
+#endif /* BACKEND_SRC_OB_HTTP_HANDLER_H_ */
